bebob-runtime: code refactoring for runtime structure
This commit refactors runtime module in runtime crate.
