+++ services/src/bin/snd-bebob-ctl-service.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2020 Takashi Sakamoto
 use snd_firewire_ctl_services::*;
-use bebob_runtime::runtime::BebobRuntime;
+use bebob_runtime::BebobRuntime;
 
 struct BebobServiceCmd;
